Ensure tests comply with verbatimModuleSyntax
Typescript [encourages][1] library authors to user
`verbatimModuleSyntax` so we have this enabled in our typescript.json.
Unfortunately not all imports have been modified in our test files.

While this may not be critical, it is still inconsistent so we clean it
up now.

This commit also rewords some of the tests to more accurately describe
what is tested.

[1]: https://www.typescriptlang.org/docs/handbook/modules/guides/choosing-compiler-options.html#im-writing-a-library