Allow to sign tokens with an incorrect key
When the issuer has multiple signing keys, the correct key should be
determined through the `kid` header of the JWT. When signing a JWT we
can now overwrite which key to choose. This allows us to create a JWT
which seems to be correctly signed, but the `kid` header will not point
to the correct key to validate the signature.